Create an actual form module
